+++ tests/sample_data/__init__.py
@@ -2,5 +2,5 @@ from pathlib import Path
 
 
 def sample_file(name: str) -> str:
-    path = (Path(__file__).parent / name).absolute()
-    return f"file://{path}"
+    path = (Path(__file__).parent / name).resolve()
+    return path.as_uri()