+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixing tests so that they can run on Windows (and still run on Linux like before) @nathanfallet
+
 ### Added
 
 ### Changed